+++ src/fortran/lat_id_driver.f90
@@ -1,31 +1,30 @@
 PROGRAM lat_id_driver
-  use lat_types
+  use niggli
   use num_types
-  use fortpy, only: pysave, fpy_read_f, fpy_read
   implicit none
 
-  real(dp) :: u_vecs(3,3), b_vecs(3,3)
-  real(dp) :: eps
-  integer :: lat_id
+  real(dp) :: lat_vecs(3,3), niggli(3,3) Nu(3,3), No(3,3), O(3,3)
+  integer :: id, Cu(3,3), Co(3,3), i
 
   character(300) :: line
 
-  eps = 1E-4
-  ! open(1,file="POSCAR",status="old")
+  open(1,file="POSCAR",status="old")
 
-  ! read(1,'(a300)') line
-  ! read(1,'(a300)') line
-  ! read(1,*) u_vecs(:,1)
-  ! read(1,*) u_vecs(:,2)
-  ! read(1,*) u_vecs(:,3)
-  ! close(1)
+  read(1,'(a300)') line
+  read(1,'(a300)') line
+  do i=1,3
+     read(1,*) lat_vecs(:,i)
+  end do
+  close(1)
 
-  call fpy_read_f("identify_lattice_vecs.in.tric2","#",u_vecs)
-  call fpy_read("identify_lattice_lat_id.out.tric2","#",lat_id)
-  print *, "lat_id",lat_id
-  call canonical_basis(lat_id,u_vecs,b_vecs)
-  call pysave(b_vecs,"canonical_basis.out.tric2")
+  call id_cell(lat_vecs,Nu,Cu,O,No,Co,id)
+  call reduce_cell(lat_vecs,niggli,Nu)
 
-  ! print *,"lat_id", lat_id
+  print *, "The cell is niggle cell number: ", id
+
+  open(2,file="Reduced",status="new")
+  do i=1,3
+     write(2,(' ','3F3.8')) niggli(:,i)
+  end do
   
 end PROGRAM lat_id_driver